llm-interaction/requirements.txt
python-dotenv
mistralai
PyYAML